The new create/delete tenant APIs on database objects can be retried if the cluster version changes and have timeouts.
