+++ fdbclient/MultiVersionAssignmentVars.h
@@ -25,104 +25,6 @@
 #include "flow/ThreadHelper.actor.h"
 
 template <class T>
-class AbortableSingleAssignmentVar final : public ThreadSingleAssignmentVar<T>, public ThreadCallback {
-public:
-	AbortableSingleAssignmentVar(ThreadFuture<T> future, ThreadFuture<Void> abortSignal)
-	  : future(future), abortSignal(abortSignal), hasBeenSet(false), callbacksCleared(false) {
-		int userParam;
-
-		ThreadSingleAssignmentVar<T>::addref();
-		ThreadSingleAssignmentVar<T>::addref();
-
-		// abortSignal comes first, because otherwise future could immediately call fire/error and attempt to remove
-		// this callback from abortSignal prematurely
-		abortSignal.callOrSetAsCallback(this, userParam, 0);
-		future.callOrSetAsCallback(this, userParam, 0);
-	}
-
-	void cancel() override {
-		cancelCallbacks();
-		ThreadSingleAssignmentVar<T>::cancel();
-	}
-
-	void cleanupUnsafe() override {
-		future.getPtr()->releaseMemory();
-		ThreadSingleAssignmentVar<T>::cleanupUnsafe();
-	}
-
-	bool canFire(int notMadeActive) const override { return true; }
-
-	void fire(const Void& unused, int& userParam) override {
-		lock.enter();
-		if (!hasBeenSet) {
-			hasBeenSet = true;
-			lock.leave();
-
-			if (future.isReady() && !future.isError()) {
-				ThreadSingleAssignmentVar<T>::send(future.get());
-			} else if (abortSignal.isReady()) {
-				ThreadSingleAssignmentVar<T>::sendError(cluster_version_changed());
-			} else {
-				ASSERT(false);
-			}
-		} else {
-			lock.leave();
-		}
-
-		cancelCallbacks();
-		ThreadSingleAssignmentVar<T>::delref();
-	}
-
-	void error(const Error& e, int& userParam) override {
-		ASSERT(future.isError());
-		lock.enter();
-		if (!hasBeenSet) {
-			hasBeenSet = true;
-			lock.leave();
-
-			ThreadSingleAssignmentVar<T>::sendError(future.getError());
-		} else {
-			lock.leave();
-		}
-
-		cancelCallbacks();
-		ThreadSingleAssignmentVar<T>::delref();
-	}
-
-private:
-	ThreadFuture<T> future;
-	ThreadFuture<Void> abortSignal;
-
-	ThreadSpinLock lock;
-	bool hasBeenSet;
-	bool callbacksCleared;
-
-	void cancelCallbacks() {
-		lock.enter();
-
-		if (!callbacksCleared) {
-			callbacksCleared = true;
-			lock.leave();
-
-			future.getPtr()->addref(); // Cancel will delref our future, but we don't want to destroy it until this
-			                           // callback gets destroyed
-			future.getPtr()->cancel();
-
-			if (abortSignal.clearCallback(this)) {
-				ThreadSingleAssignmentVar<T>::delref();
-			}
-		} else {
-			lock.leave();
-		}
-	}
-};
-
-template <class T>
-ThreadFuture<T> abortableFuture(ThreadFuture<T> f, ThreadFuture<Void> abortSignal) {
-	return ThreadFuture<T>(new AbortableSingleAssignmentVar<T>(f, abortSignal));
-}
-
-template <class T>
 class DLThreadSingleAssignmentVar final : public ThreadSingleAssignmentVar<T> {
 public:
 	DLThreadSingleAssignmentVar(Reference<FdbCApi> api,
@@ -222,173 +124,6 @@ ThreadFuture<T> toThreadFuture(Reference<FdbCApi> api,
                                FdbCApi::FDBFuture* f,
                                std::function<T(FdbCApi::FDBFuture* f, FdbCApi* api)> extractValue) {
 	return ThreadFuture<T>(new DLThreadSingleAssignmentVar<T>(api, f, extractValue));
-}
-
-template <class S, class T>
-class MapSingleAssignmentVar final : public ThreadSingleAssignmentVar<T>, ThreadCallback {
-public:
-	MapSingleAssignmentVar(ThreadFuture<S> source, std::function<ErrorOr<T>(ErrorOr<S>)> mapValue)
-	  : source(source), mapValue(mapValue) {
-		ThreadSingleAssignmentVar<T>::addref();
-
-		int userParam;
-		source.callOrSetAsCallback(this, userParam, 0);
-	}
-
-	void cancel() override {
-		source.getPtr()->addref(); // Cancel will delref our future, but we don't want to destroy it until this callback
-		                           // gets destroyed
-		source.getPtr()->cancel();
-		ThreadSingleAssignmentVar<T>::cancel();
-	}
-
-	void cleanupUnsafe() override {
-		source.getPtr()->releaseMemory();
-		ThreadSingleAssignmentVar<T>::cleanupUnsafe();
-	}
-
-	bool canFire(int notMadeActive) const override { return true; }
-
-	void fire(const Void& unused, int& userParam) override {
-		sendResult(mapValue(source.get()));
-		ThreadSingleAssignmentVar<T>::delref();
-	}
-
-	void error(const Error& e, int& userParam) override {
-		sendResult(mapValue(source.getError()));
-		ThreadSingleAssignmentVar<T>::delref();
-	}
-
-private:
-	ThreadFuture<S> source;
-	const std::function<ErrorOr<T>(ErrorOr<S>)> mapValue;
-
-	void sendResult(ErrorOr<T> result) {
-		if (result.isError()) {
-			ThreadSingleAssignmentVar<T>::sendError(result.getError());
-		} else {
-			ThreadSingleAssignmentVar<T>::send(result.get());
-		}
-	}
-};
-
-template <class S, class T>
-ThreadFuture<T> mapThreadFuture(ThreadFuture<S> source, std::function<ErrorOr<T>(ErrorOr<S>)> mapValue) {
-	return ThreadFuture<T>(new MapSingleAssignmentVar<S, T>(source, mapValue));
-}
-
-template <class S, class T>
-class FlatMapSingleAssignmentVar final : public ThreadSingleAssignmentVar<T>, ThreadCallback {
-public:
-	FlatMapSingleAssignmentVar(ThreadFuture<S> source, std::function<ErrorOr<ThreadFuture<T>>(ErrorOr<S>)> mapValue)
-	  : source(source), cancelled(false), released(false), mapValue(mapValue) {
-		ThreadSingleAssignmentVar<T>::addref();
-
-		int userParam;
-		source.callOrSetAsCallback(this, userParam, 0);
-	}
-
-	void cancel() override {
-		source.getPtr()->addref(); // Cancel will delref our future, but we don't want to destroy it until this callback
-		                           // gets destroyed
-		source.getPtr()->cancel();
-
-		lock.enter();
-		cancelled = true;
-		if (mappedFuture.isValid()) {
-			lock.leave();
-			mappedFuture.getPtr()->addref();
-			mappedFuture.getPtr()->cancel();
-		} else {
-			lock.leave();
-		}
-
-		ThreadSingleAssignmentVar<T>::cancel();
-	}
-
-	void cleanupUnsafe() override {
-		source.getPtr()->releaseMemory();
-
-		lock.enter();
-		released = true;
-		if (mappedFuture.isValid()) {
-			lock.leave();
-			mappedFuture.getPtr()->releaseMemory();
-		} else {
-			lock.leave();
-		}
-
-		ThreadSingleAssignmentVar<T>::cleanupUnsafe();
-	}
-
-	bool canFire(int notMadeActive) const override { return true; }
-
-	void fire(const Void& unused, int& userParam) override {
-		if (mappedFuture.isValid()) {
-			sendResult(mappedFuture.get());
-		} else {
-			setMappedFuture(mapValue(source.get()));
-		}
-
-		ThreadSingleAssignmentVar<T>::delref();
-	}
-
-	void error(const Error& e, int& userParam) override {
-		if (mappedFuture.isValid()) {
-			sendResult(mappedFuture.getError());
-		} else {
-			setMappedFuture(mapValue(source.getError()));
-		}
-
-		ThreadSingleAssignmentVar<T>::delref();
-	}
-
-private:
-	ThreadFuture<S> source;
-	ThreadFuture<T> mappedFuture;
-	bool cancelled;
-	bool released;
-	const std::function<ErrorOr<ThreadFuture<T>>(ErrorOr<S>)> mapValue;
-
-	ThreadSpinLock lock;
-
-	void setMappedFuture(ErrorOr<ThreadFuture<T>> f) {
-		if (f.isError()) {
-			sendResult(f.getError());
-		} else {
-			lock.enter();
-			mappedFuture = f.get();
-			bool doCancel = cancelled;
-			bool doRelease = released;
-			lock.leave();
-
-			if (doCancel) {
-				mappedFuture.getPtr()->addref();
-				mappedFuture.getPtr()->cancel();
-			}
-			if (doRelease) {
-				mappedFuture.getPtr()->releaseMemory();
-			}
-
-			int userParam;
-			ThreadSingleAssignmentVar<T>::addref();
-			mappedFuture.callOrSetAsCallback(this, userParam, 0);
-		}
-	}
-
-	void sendResult(ErrorOr<T> result) {
-		if (result.isError()) {
-			ThreadSingleAssignmentVar<T>::sendError(result.getError());
-		} else {
-			ThreadSingleAssignmentVar<T>::send(result.get());
-		}
-	}
-};
-
-template <class S, class T>
-ThreadFuture<T> flatMapThreadFuture(ThreadFuture<S> source,
-                                    std::function<ErrorOr<ThreadFuture<T>>(ErrorOr<S>)> mapValue) {
-	return ThreadFuture<T>(new FlatMapSingleAssignmentVar<S, T>(source, mapValue));
 }
 
 #endif
